pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "rag-knowledge-assistant-mlo-project"
version = "0.1.0"
description = ""
requires-python = ">=3.11"

dependencies = [
  "faiss-cpu>=1.13.2",
  "fastapi>=0.128.0",
  "numpy>=2.4.1",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "pypdf>=6.6.0",
  "rich>=14.2.0",
  "sentence-transformers>=5.2.0",
  "uvicorn>=0.40.0",
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
